refactor(il/io): split parser implementation
